pretrain/requirements.txt
numpy>=1.24
pandas>=2.1
scipy>=1.11
scikit-learn>=1.3
torch>=2.1
pyyaml>=6.0
requests>=2.31
tqdm>=4.66
einops>=0.7
